aws_kb/pyproject.toml
[project]
name = "aws-kb-mcp"
version = "1.0.0"
description = "AWS Knowledge Base MCP Server for SDLC Documentation"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.1.0",
    "boto3>=1.34.0",
]
